Internal #861: Refactor Window Aggregates
Move the window Aggregate function classes to separate files.
